Logic changes, some tests implemented, actuator added.

+++ src/main/java/faang/school/projectservice/service/team/TeamMemberService.java
@@ -1,6 +1,7 @@
 package faang.school.projectservice.service.team;
 
 import faang.school.projectservice.client.UserServiceClient;
+import faang.school.projectservice.config.context.UserContext;
 import faang.school.projectservice.dto.client.UserDto;
 import faang.school.projectservice.model.Project;
 import faang.school.projectservice.model.Team;
@@ -9,6 +10,7 @@ import faang.school.projectservice.model.TeamRole;
 import faang.school.projectservice.repository.ProjectRepository;
 import faang.school.projectservice.repository.TeamMemberRepository;
 import faang.school.projectservice.repository.TeamRepository;
+import faang.school.projectservice.util.TeamMemberUtil;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,7 +18,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 
@@ -28,17 +29,23 @@ public class TeamMemberService {
     private final TeamRepository teamRepository;
     private final UserServiceClient userServiceClient;
     private final ProjectRepository projectRepository;
+    private final UserContext userContext;
+
+    @Transactional
+    public List<TeamMember> getMembersForTeam(Long teamId) {
+        Team target = teamRepository.findById(teamId).orElseThrow(
+                () -> new EntityNotFoundException(String.format("No team found for ID = %s", teamId)));
+        return target.getTeamMembers();
+    }
 
     @Transactional
     public List<TeamMember> addToTeam(Long teamId, TeamRole role, List<Long> userIds) {
         Team toReceiveMembers = teamRepository.findById(teamId).orElseThrow(
-                () -> new EntityNotFoundException("Team with id " + teamId + " not found"));
+                () -> new EntityNotFoundException(String.format("No team found for ID = %s", teamId)));
         Project teamProject = toReceiveMembers.getProject();
-        TeamMember caller = teamMemberRepository.findByUserIdAndProjectId(null, teamProject.getId())
-                .orElseThrow(() -> new IllegalStateException("Caller not registered for this project"));
-        validateCallerIsTeamLead(caller);
-        List<TeamMember> verifiedMembers = getMockedUsers(userIds)
-//        List<TeamMember> verifiedMembers = userServiceClient.getUsersByIds(userIds)
+        TeamMember caller = teamMemberRepository.findByUserIdAndProjectId(userContext.getUserId(), teamProject.getId());
+        TeamMemberUtil.validateOwnerOrTeamLead(teamProject, userContext.getUserId(), caller);
+        List<TeamMember> verifiedMembers = userServiceClient.getUsersByIds(userIds)
                 .stream()
                 .map(dto -> verifyMember(dto, teamProject.getId(), toReceiveMembers, role))
                 .toList();
@@ -48,50 +55,37 @@ public class TeamMemberService {
     }
 
     @Transactional
-    public TeamMember updateMemberRoles(Long teamMemberId, List<TeamRole> roles) {
+    public TeamMember updateMemberRoles(Long teamMemberId, List<TeamRole> newRoles) {
         TeamMember toUpdate = teamMemberRepository.findByIdOrThrow(teamMemberId);
-        Project contributingTo = toUpdate.getTeam().getProject();
-        // x-user-id
-        TeamMember caller = teamMemberRepository.findByUserIdAndProjectId(null, contributingTo.getId()).orElseThrow(
-                () -> {
-                    String message = String.format("No member found for user ID = %s and project ID = %s", null, contributingTo.getId());
-                    return new IllegalStateException(message);
-                });
-        validateCallerIsTeamLead(caller);
-        if (roles.isEmpty()) {
-            throw new IllegalArgumentException("Team member must have at least one role");
-        }
-        List<TeamRole> updatedRoles = toUpdate.getRoles().stream()
-                        .filter(roles::contains)
-                        .toList();
-        toUpdate.setRoles(updatedRoles);
+        Project teamProject = toUpdate.getTeam().getProject();
+        TeamMember caller = teamMemberRepository.findByUserIdAndProjectId(userContext.getUserId(), teamProject.getId());
+        TeamMemberUtil.validateTeamLead(caller);
+        TeamMemberUtil.validateNewRolesNotEmpty(newRoles);
+        toUpdate.setRoles(newRoles);
         return teamMemberRepository.save(toUpdate);
     }
 
     @Transactional
     public TeamMember updateMemberNickname(Long teamMemberId, String nickname) {
         TeamMember toUpdate = teamMemberRepository.findByIdOrThrow(teamMemberId);
-        if (!toUpdate.getUserId().equals(null)) {
-            throw new SecurityException("Calling user is not a member to be updated");
-        }
+        TeamMemberUtil.validateCallerOwnsAccount(toUpdate, userContext.getUserId());
         toUpdate.setNickname(nickname);
         return teamMemberRepository.save(toUpdate);
     }
 
     @Transactional
     public void removeFromProject(Long projectId, List<Long> userIds) {
-        TeamMember caller = teamMemberRepository.findByUserIdAndProjectId(null, projectId).orElseThrow(
-                () -> new IllegalStateException("Calling user is not registered for this project"));
-        validateCallerIsTeamLead(caller);
-        List<TeamMember> membersToRemove = userIds.stream()
-                .map(userId -> teamMemberRepository.findByUserIdAndProjectId(projectId, userId)
-                        .orElseThrow(() -> new EntityNotFoundException("Team member with id " + userId + " not found")))
+        Project toBeRemovedFrom = projectRepository.findByIdThrowing(projectId);
+        TeamMemberUtil.validateProjectOwner(toBeRemovedFrom, userContext.getUserId());
+        List<TeamMember> membersToRemove = userServiceClient.getUsersByIds(userIds).stream()
+                .map(UserDto::getId)
+                .map(userId -> teamMemberRepository.findByUserIdAndProjectId(projectId, userId))
                 .toList();
         teamMemberRepository.deleteAll(membersToRemove);
     }
 
     @Transactional(readOnly = true)
-    public List<TeamMember> getFilteredMembers(Long projectId, String nickname, TeamRole role) {
+    public List<TeamMember> getProjectMembersFiltered(Long projectId, String nickname, TeamRole role) {
         Project target = projectRepository.getProjectByIdOrThrow(projectId);
         Stream<TeamMember> stream = target.getTeams().stream()
                 .flatMap(team -> team.getTeamMembers().stream());
@@ -108,14 +102,14 @@ public class TeamMemberService {
     }
 
     @Transactional(readOnly = true)
-    public List<TeamMember> getMembersWithSameUser(Long userId) {
+    public List<TeamMember> getMembersForUser(Long userId) {
         UserDto validated = userServiceClient.getUser(userId);
         return teamMemberRepository.findByUserId(validated.getId());
     }
 
     private TeamMember verifyMember(UserDto userData, Long projectId, Team target, TeamRole role) {
-        Optional<TeamMember> optional = teamMemberRepository.findByUserIdAndProjectId(userData.getId(), projectId);
-        if (optional.isEmpty()) {
+        TeamMember member = teamMemberRepository.findByUserIdAndProjectId(userData.getId(), projectId);
+        if (Objects.isNull(member)) {
             return TeamMember.builder()
                     .userId(userData.getId())
                     .nickname(userData.getUsername())
@@ -123,35 +117,18 @@ public class TeamMemberService {
                     .roles(List.of(role))
                     .build();
         }
-        TeamMember existing = optional.get();
-        if (existing.getRoles().contains(role)) {
+        if (member.getRoles().contains(role)) {
             throw new IllegalStateException("Team member with user ID" + userData.getId() + "already has this role on the project.");
         }
-        existing.getRoles().add(role);
-        return existing;
+        member.getRoles().add(role);
+        return member;
     }
 
     private Predicate<TeamMember> nicknameFilter(String filter) {
-        return member -> member.getNickname().equals(filter);
+        return member -> member.getNickname().contains(filter);
     }
 
     private Predicate<TeamMember> roleFilter(TeamRole filter) {
         return member -> member.getRoles().contains(filter);
-    }
-
-    private void validateCallerIsTeamLead(TeamMember caller) {
-        if (!caller.getRoles().contains(TeamRole.TEAM_LEAD)) {
-            throw new SecurityException("Caller does not have team lead privileges");
-        }
-    }
-
-    private List<UserDto> getMockedUsers(List<Long> userIds) {
-        return userIds.stream()
-                .map(l -> UserDto.builder()
-                        .id(l)
-                        .username("member_fuck_knows_who_" + l)
-                        .email(l + "@fuck-you.com")
-                        .build())
-                .toList();
     }
 }
